chore: Move unrelated functions away from expressions.ts
