+++ packages/@cdktf/hcl2cdk/lib/utils.ts
@@ -4,6 +4,8 @@ import camelcase from "camelcase";
 
 export { logger } from "@cdktf/commons";
 
+export const leaveCommentText = `Please leave a comment at https://cdk.tf/bugs/convert-expressions if you run into this issue.`;
+
 export const camelCase = (str: string) => camelcase(str.replace(/[-/]/g, "_"));
 export const pascalCase = (str: string) =>
   camelcase(str.replace(/[-/]/g, "_"), { pascalCase: true });
